requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.0
billiard==4.2.1
CacheControl==0.14.0
cachetools==5.5.0
celery==5.3.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
django-modeltranslation==0.19.10
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.8
firebase-admin==6.5.0
flower==2.0.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
msgpack==1.1.0
packaging==24.1
pillow==11.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.2
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
